requirements.txt
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=1.0.0
shiny>=1.0.0
shinywidgets>=0.3.0
plotly>=5.0.0
psycopg2-binary>=2.9.0
folium>=0.14.0
